+++ pkg/providers/cloudcertificates/testdata/TestResCertificate/create/min_contiguous_us.tf
@@ -3,16 +3,11 @@ provider "akamai" {
 }
 
 resource "akamai_cloudcertificates_certificate" "test" {
-  base_name      = "test-name"
   contract_id    = "test_contract"
   group_id       = "123"
-  key_size       = "P-256"
-  key_type       = "ECDSA"
+  key_size       = "2048"
+  key_type       = "RSA"
   secure_network = "ENHANCED_TLS"
   geo_class      = "CONTIGUOUS_US"
-  sans           = ["test.example.com", "test.example2.com"]
-  subject = {
-    state    = "CA"
-    locality = "Test City"
-  }
+  sans           = ["test.example.com"]
 }